+++ Trainer_v4/Constants.cs
@@ -1,29 +1,31 @@
 ﻿namespace Trainer_v4
 {
-    public static class Constants
-    {
-        public static readonly int SETTINGS_WINDOW_SKIP_ROWS = 7;
+	public static class Constants
+	{
+		public static readonly int SETTINGS_WINDOW_SKIP_ROWS = 7;
 
-        public static readonly int ELEMENT_WIDTH = 150;
-        public static readonly int ELEMENT_HEIGHT = 32;
+		public static readonly int ELEMENT_WIDTH = 150;
+		public static readonly int ELEMENT_HEIGHT = 32;
 
-        public static readonly int Y_SETTINGS_WINDOW_OFFSET = 260;
-        public static readonly int X_SETTINGS_WINDOW = 670;
+		public static readonly int Y_SETTINGS_WINDOW_OFFSET = 260;
+		public static readonly int X_SETTINGS_WINDOW = 831;
 
-        public static readonly int Y_EMPLOYEESKILLCHANGE_WINDOW_OFFSET = 64;
-        public static readonly int X_EMPLOYEESKILLCHANGE_WINDOW = 300;
+		public static readonly int Y_EMPLOYEESKILLCHANGE_WINDOW_OFFSET = 64;
+		public static readonly int X_EMPLOYEESKILLCHANGE_WINDOW = 300;
 
-        public static readonly int FIRST_COLUMN = 1;
-        public static readonly int SECOND_COLUMN = 161;
-        public static readonly int THIRD_COLUMN = 322;
-        public static readonly int FOURTH_COLUMN = 483;
+		public static readonly int FIRST_COLUMN = 1;
+		public static readonly int SECOND_COLUMN = 161;
+		public static readonly int THIRD_COLUMN = 322;
+		public static readonly int FOURTH_COLUMN = 483;
+		public static readonly int FIFTH_COLUMN = 644;
 
-        public static readonly int FIRST_ROW;
-        public static readonly int SECOND_ROW = 32;
-        public static readonly int THIRD_ROW = 64;
-        public static readonly int FOURTH_ROW = 96;
-        public static readonly int FIFTH_ROW = 128;
-        public static readonly int SIXTH_ROW = 160;
-        public static readonly int SEVENTH_ROW = 192;
-    }
+		public static readonly int FIRST_ROW;
+		public static readonly int SECOND_ROW = 32;
+		public static readonly int THIRD_ROW = 64;
+		public static readonly int FOURTH_ROW = 96;
+		public static readonly int FIFTH_ROW = 128;
+		public static readonly int SIXTH_ROW = 160;
+		public static readonly int SEVENTH_ROW = 192;
+		public static readonly int EIGHTH_ROW = 224;
+	}
 }