+++ Trainer_v4/PropertyHelper.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using OrbCreationExtensions;
 
 namespace Trainer_v4
 {
@@ -9,9 +10,6 @@ namespace Trainer_v4
 		public static bool RewardIsGained { get; set; }
 		public static bool DealIsPushed { get; set; }
 		public static string ProductPriceName { get; set; }
-
-		private static int _UltraEfficiencyStore = 20;
-		public static int UltraEfficiencyMultipplier { get { return _UltraEfficiencyStore; } set { _UltraEfficiencyStore = value > 20 ? value : 20; } }
 
 		public static bool IsGameLoaded
 		{
@@ -28,6 +26,22 @@ namespace Trainer_v4
 			get { return new[] { 1f, 0.75f, 0.5f }; }
 		}
 
+		public static List<KeyValuePair<string, object>> EfficiencyValues
+		{
+			get
+			{
+				return new List<KeyValuePair<string, object>>
+				{
+					new KeyValuePair<string, object>("100%", 1),
+					new KeyValuePair<string, object>("200%", 2),
+					new KeyValuePair<string, object>("500%", 5),
+					new KeyValuePair<string, object>("1000%", 10),
+					new KeyValuePair<string, object>("2000%", 20),
+					new KeyValuePair<string, object>("4000%", 40)
+				};
+			}
+		}
+
 		private static Dictionary<string, bool> _settings = new Dictionary<string, bool>
 		{
 			{"NoStress", false},
@@ -40,7 +54,6 @@ namespace Trainer_v4
 			{"TemperatureLock", false},
 			{"NoWaterElectricity", false},
 			{"NoNeeds", false},
-			{"FullEfficiency", false},
 			{"FreeEmployees", false},
 			{"LockAge", false},
 			{"AutoDistributionDeals", false},
@@ -52,7 +65,6 @@ namespace Trainer_v4
 			{"IncreaseBookshelfSkill", false},
 			{"NoMaintenance", false},
 			{"NoSickness", false},
-			{"UltraEfficiency", false},
 			{"FullSatisfaction", false},
 			{"DisableSkillDecay", false},
 			{"DisableBurglars", false},
@@ -76,6 +88,12 @@ namespace Trainer_v4
 			{"Designer", false}
 		};
 
+		private static Dictionary<string, object> _stores = new Dictionary<string, object>
+		{
+			{"EfficiencyStore", 2},
+			{"LeadEfficiencyStore", 4}
+		};
+
 		public static Dictionary<string, bool> Settings
 		{
 			get { return _settings; }
@@ -84,6 +102,11 @@ namespace Trainer_v4
 		public static Dictionary<string, bool> RolesList
 		{
 			get { return _rolesList; }
+		}
+
+		public static Dictionary<string, object> Stores
+		{
+			get { return _stores; }
 		}
 
 		public static Dictionary<string, bool> SpecializationsList { get; set; }
@@ -95,15 +118,44 @@ namespace Trainer_v4
 			{
 				return value;
 			}
-			else
+			return false;
+		}
+
+		public static object GetProperty(Dictionary<string, object> properties, string key)
+		{
+			object value;
+			if (properties.TryGetValue(key, out value))
 			{
-				return false;
+				return value;
 			}
+			return null;
 		}
 
 		public static void SetProperty(Dictionary<string, bool> properties, string key, bool value)
 		{
 			properties[key] = value;
+		}
+
+		public static void SetProperty(Dictionary<string, object> properties, string key, object value)
+		{
+			properties[key] = value;
+		}
+
+		public static int GetIndex(List<KeyValuePair<string, object>> values, Dictionary<string, object> properties, string store, int valueType)
+		{
+			switch (valueType)
+			{
+				case 1:
+					return values.FindIndex(x => x.Value.MakeInt() == GetProperty(properties, store).MakeInt());
+				case 2:
+					return values.FindIndex(x => x.Value.MakeFloat() == GetProperty(properties, store).MakeFloat());
+				case 3:
+					return values.FindIndex(x => x.Value.MakeString() == GetProperty(properties, store).MakeString());
+				case 4:
+					return values.FindIndex(x => x.Value.MakeBool() == GetProperty(properties, store).MakeBool());
+				default:
+					throw new NotImplementedException("Method GetIndex received an unkown value type as parameter");
+			}
 		}
 
 		#region extensions